requirements.txt
APScheduler==3.11.0
asgiref==3.8.1
attrs==24.3.0
autobahn==24.4.2
Automat==24.8.1
cffi==1.17.1
channels==4.2.0
channels_redis==4.2.1
colorama==0.4.6
constantly==23.10.4
cryptography==44.0.0
daphne==4.1.2
Django==5.1.4
django-cors-headers==4.6.0
django-environ==0.12.0
django-redis==5.4.0
djangorestframework==3.15.2
gunicorn==23.0.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
jsonpickle==4.0.1
msgpack==1.1.0
packaging==24.2
pillow==11.1.0
psycopg2==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pyOpenSSL==24.3.0
qrcode==8.0
redis==5.2.1
service-identity==24.2.0
setuptools==75.6.0
sqlparse==0.5.3
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
zope.interface==7.2
